♻️ refactor : toast style 을 상수화해서 constants 파일로 저장, toastStore 코드 수정((#23)
